requirements.txt
requests
selenium
beautifulsoup4
lmdb
numpy
Pillow
six
opencv-python
PyYAML
ipython
easydict
torch
torchvision
torchaudio
tqdm
tensorboard
matplotlib
editdistance
natsort
nltk